+++ deployment/aws/cdk/app.py
@@ -3,21 +3,22 @@
 import os
 from typing import Any, Dict, List, Optional, Union
 
-from aws_cdk import aws_apigatewayv2 as apigw
-from aws_cdk import aws_apigatewayv2_integrations as apigw_integrations
+from aws_cdk import App, CfnOutput, Duration, Stack, Tags
+from aws_cdk import aws_apigatewayv2_alpha as apigw
 from aws_cdk import aws_ec2 as ec2
 from aws_cdk import aws_ecs as ecs
 from aws_cdk import aws_ecs_patterns as ecs_patterns
 from aws_cdk import aws_iam as iam
 from aws_cdk import aws_lambda
 from aws_cdk import aws_logs as logs
-from aws_cdk import core
+from aws_cdk.aws_apigatewayv2_integrations_alpha import HttpLambdaIntegration
 from config import StackSettings
+from constructs import Construct
 
 settings = StackSettings()
 
 
-class titilerLambdaStack(core.Stack):
+class titilerLambdaStack(Stack):
     """
     Titiler Lambda Stack
 
@@ -29,11 +30,11 @@ class titilerLambdaStack(core.Stack):
 
     def __init__(
         self,
-        scope: core.Construct,
+        scope: Construct,
         id: str,
         memory: int = 1024,
         timeout: int = 30,
-        runtime: aws_lambda.Runtime = aws_lambda.Runtime.PYTHON_3_9,
+        runtime: aws_lambda.Runtime = aws_lambda.Runtime.PYTHON_3_10,
         concurrent: Optional[int] = None,
         permissions: Optional[List[iam.PolicyStatement]] = None,
         environment: Optional[Dict] = None,
@@ -57,7 +58,7 @@ class titilerLambdaStack(core.Stack):
             handler="handler.handler",
             memory_size=memory,
             reserved_concurrent_executions=concurrent,
-            timeout=core.Duration.seconds(timeout),
+            timeout=Duration.seconds(timeout),
             environment=environment,
             log_retention=logs.RetentionDays.ONE_WEEK,
         )
@@ -68,19 +69,19 @@ class titilerLambdaStack(core.Stack):
         api = apigw.HttpApi(
             self,
             f"{id}-endpoint",
-            default_integration=apigw_integrations.HttpLambdaIntegration(
+            default_integration=HttpLambdaIntegration(
                 f"{id}-integration", handler=lambda_function
             ),
         )
-        core.CfnOutput(self, "Endpoint", value=api.url)
+        CfnOutput(self, "Endpoint", value=api.url)
 
 
-class titilerECSStack(core.Stack):
+class titilerECSStack(Stack):
     """Titiler ECS Fargate Stack."""
 
     def __init__(
         self,
-        scope: core.Construct,
+        scope: Construct,
         id: str,
         cpu: Union[int, float] = 256,
         memory: Union[int, float] = 512,
@@ -123,7 +124,7 @@ class titilerECSStack(core.Stack):
             listener_port=80,
             task_image_options=ecs_patterns.ApplicationLoadBalancedTaskImageOptions(
                 image=ecs.ContainerImage.from_registry(
-                    f"public.ecr.aws/developmentseed/titiler:{settings.image_version}",
+                    f"ghcr.io/developmentseed/titiler:{settings.image_version}",
                 ),
                 container_port=80,
                 environment=task_env,
@@ -142,8 +143,8 @@ class titilerECSStack(core.Stack):
         scalable_target.scale_on_request_count(
             "RequestScaling",
             requests_per_target=50,
-            scale_in_cooldown=core.Duration.seconds(240),
-            scale_out_cooldown=core.Duration.seconds(30),
+            scale_in_cooldown=Duration.seconds(240),
+            scale_out_cooldown=Duration.seconds(30),
             target_group=fargate_service.target_group,
         )
 
@@ -161,7 +162,7 @@ class titilerECSStack(core.Stack):
         )
 
 
-app = core.App()
+app = App()
 
 perms = []
 if settings.buckets:
@@ -175,20 +176,9 @@ if settings.buckets:
     )
 
 
-# Tag infrastructure
-for key, value in {
-    "Project": settings.name,
-    "Stack": settings.stage,
-    "Owner": settings.owner,
-    "Client": settings.client,
-}.items():
-    if value:
-        core.Tag.add(app, key, value)
-
-ecs_stackname = f"{settings.name}-ecs-{settings.stage}"
-titilerECSStack(
+ecs_stack = titilerECSStack(
     app,
-    ecs_stackname,
+    f"{settings.name}-ecs-{settings.stage}",
     cpu=settings.task_cpu,
     memory=settings.task_memory,
     mincount=settings.min_ecs_instances,
@@ -197,15 +187,26 @@ titilerECSStack(
     environment=settings.env,
 )
 
-lambda_stackname = f"{settings.name}-lambda-{settings.stage}"
-titilerLambdaStack(
+lambda_stack = titilerLambdaStack(
     app,
-    lambda_stackname,
+    f"{settings.name}-lambda-{settings.stage}",
     memory=settings.memory,
     timeout=settings.timeout,
     concurrent=settings.max_concurrent,
     permissions=perms,
     environment=settings.env,
 )
+
+# Tag infrastructure
+for key, value in {
+    "Project": settings.name,
+    "Stack": settings.stage,
+    "Owner": settings.owner,
+    "Client": settings.client,
+}.items():
+    if value:
+        Tags.of(ecs_stack).add(key, value)
+        Tags.of(lambda_stack).add(key, value)
+
 
 app.synth()