+++ src/titiler/application/tests/routes/test_cog.py
@@ -59,7 +59,7 @@ def test_wmts(rio, app):
     assert response.headers["content-type"] == "application/xml"
     assert response.headers["Cache-Control"] == "private, max-age=3600"
     assert (
-        "http://testserver/cog/WMTSCapabilities.xml?url=https://myurl.com/cog.tif"
+        "http://testserver/cog/WMTSCapabilities.xml?url=https"
         in response.content.decode()
     )
     assert "<ows:Identifier>cogeo</ows:Identifier>" in response.content.decode()